requirements.txt
asgiref==3.5.0
attrdict==2.0.1
cached-property==1.5.2
certifi==2021.10.8
configparser==5.2.0
dataclasses==0.6
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
minio==6.0.2
PyJWT==2.3.0
pyminio==0.2.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
psycopg2